Mobile: add new mobile-view style to make look views the same as dialogs
Also, setting the headerVisible flag to true on a view now removes
the tab. Having tabs on a mobile device is confusing. This is the way
to remove them.
This makes it possible to style a view like a dialog which has the
advantage that it does not matter whether a form is opened as dialog
or as view on a mobile device, they look the same.

The previous behavior can easily be reverted by disabling the new
transformation USE_DIALOG_STYLE_FOR_VIEW.