Widget: optimize animate removal (POC)
Currently, if a form with animateRemoval = true
is removed (using formHide) an exception could occur parts
of the form should be rendered while it is being removed.
Normally this cannot happen because no events will be fired for
a destroyed form. But since it is only removed not destroyed events
can occur.
The change tries to disable all render operations on a widget that
is being removed. Open task: We should test the performance impact
because now every call to this.rendered traverses all parents.
Maybe it would be better to set the property removalPending on every
child node so that this.rendered is faster.