Data object signature test: add initial contribution
The data object signature test creates a signature file representing the structure (e.g. attributes with their types) of data objects annotated with a type version. Any changes in such a data object will result in a test failure to notify the developer that if such a data object persisted, a migration might be necessary.